pyproject.toml
[project]
name = "bom-radar-scraper"
version = "0.1.0"
description = "A script to scrape and process BOM radar images."
requires-python = ">=3.8"
dependencies = [
    "requests",
    "Pillow",
    "pytz",
]

[tool.uv]
dev-dependencies = []
